Removed JS files to consolidate 3 components into 1.
