+++ force-app/main/default/lwc/participantSelector/participantSelector.css
