--- force-app/main/default/lwc/selectParticipantContainer/selectParticipantContainer.html
@@ -1,17 +1,0 @@
-<template>
-    <c-section title={sectionHeader}></c-section>
-
-    <lightning-layout class="container-height">
-        <lightning-layout-item size="8">
-            <c-participant-selector
-                onselectparticipants={handleSelectedParticipants}
-            ></c-participant-selector>
-        </lightning-layout-item>
-        <lightning-layout-item size="4">
-            <c-selected-participants
-                selected-participants={selectedParticipants}
-                ondeletedparticipant={handleDeletedParticipant}
-            ></c-selected-participants>
-        </lightning-layout-item>
-    </lightning-layout>
-</template>
