--- force-app/main/default/lwc/selectParticipantContainer/selectParticipantContainer.js-meta.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata">
-    <apiVersion>49.0</apiVersion>
-    <isExposed>true</isExposed>
-    <masterLabel>Select Participant</masterLabel>
-    <targets>
-        <target>lightning__AppPage</target>
-        <target>lightning__RecordPage</target>
-    </targets>
-</LightningComponentBundle>
