--- force-app/main/default/lwc/selectedParticipants/selectedParticipants.html
@@ -1,32 +1,0 @@
-<template>
-    <lightning-layout>
-        <lightning-layout-item size="12" class="slds-var-m-top_medium">
-            <div class="slds-box slds-p-around_none  slds-var-m-top_medium">
-                <div
-                    class="slds-align_absolute-center slds-var-p-around_x-small slds-text-heading_small"
-                >
-                    {scheduleHeader}
-                </div>
-                <template if:false={noRecordsSelected}>
-                    <div style="width:fit-content">
-                        <lightning-datatable
-                            columns={columns}
-                            key-field="Id"
-                            hide-checkbox-column
-                            onrowaction={handleRemove}
-                            data={selectedParticipants}
-                            resize-column-disabled
-                        ></lightning-datatable>
-                    </div>
-                </template>
-                <template if:true={noRecordsSelected}>
-                    <div
-                        class="width:fit-content slds-align_absolute-center slds-var-p-around_small slds-text-heading_small slds-border_top"
-                    >
-                        {labels.noRecordsSelected}
-                    </div>
-                </template>
-            </div>
-        </lightning-layout-item>
-    </lightning-layout>
-</template>
