+++ force-app/main/default/lwc/serviceScheduleCreator/serviceScheduleCreator.html
@@ -12,7 +12,7 @@
             <c-review-sessions></c-review-sessions>
         </template>
         <template if:true={showStep3}>
-            <!-- Add Participants -->
+            <c-participant-selector service-id={serviceId}></c-participant-selector>
         </template>
         <template if:true={showStep4}>
             <c-service-schedule-review></c-service-schedule-review>
